已修改完毕，数据库也基本完善，准备开始开发具体业务。

+++ zsblogs/src/main/java/com/zs/service/PerSer.java
@@ -1,10 +1,9 @@
 package com.zs.service;
 
-import java.math.BigDecimal;
 
 import com.zs.entity.Permission;
 
-public interface PerSer extends BaseService<Permission, BigDecimal>{
+public interface PerSer extends BaseService<Permission, Integer>{
 
 	public Permission get(String url,String method);
 }
