+++ zsblogs/src/main/java/com/zs/service/TimeLineSer.java
@@ -1,9 +1,7 @@
 package com.zs.service;
 
-import java.math.BigDecimal;
-
 import com.zs.entity.Timeline;
 
-public interface TimeLineSer extends BaseService<Timeline, BigDecimal>{
+public interface TimeLineSer extends BaseService<Timeline, Integer>{
 
 }
